+++ noodles-csi/src/async/io.rs
@@ -1,0 +1,5 @@
+//! Async CSI.
+
+mod reader;
+
+pub use self::reader::Reader;